requirements.txt
Django 
pyzmq
django-bootstrap3
django.contrib.postgres
djangorestframework==3.8.2
django.core.urlresolvers